requirements.txt
# cyber-pi: Enterprise Threat Intelligence Platform
# Zero-budget, massive parallel processing with GPU acceleration

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# ============================================================================
# ASYNC & PARALLEL PROCESSING
# ============================================================================
aiohttp>=3.11.0
asyncio>=3.4.3
aiodns>=3.2.0
aiofiles>=24.1.0

# ============================================================================
# DATA COLLECTION
# ============================================================================
# RSS & Web Scraping
feedparser>=6.0.11
beautifulsoup4>=4.12.3
lxml>=5.3.0
scrapy>=2.11.2
selenium>=4.26.1
requests>=2.32.3

# Social Media APIs
tweepy>=4.14.0          # Twitter/X
praw>=7.8.1             # Reddit
PyGithub>=2.5.0         # GitHub

# Content Extraction
newspaper3k>=0.2.8
readability-lxml>=0.8.1
html2text>=2024.2.26

# ============================================================================
# GPU-ACCELERATED NLP & ML
# ============================================================================
torch>=2.5.1
transformers>=4.46.3
sentence-transformers>=3.3.1
spacy>=3.8.2
nltk>=3.9.1

# ============================================================================
# TRI-MODAL DATABASE CLIENTS
# ============================================================================
# Redis
redis[hiredis]>=5.2.0
redis-om>=0.3.2

# Neo4j
neo4j>=5.26.0

# Weaviate
weaviate-client>=4.9.3

# PostgreSQL (optional structured storage)
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36

# ============================================================================
# DATA PROCESSING & ANALYSIS
# ============================================================================
pandas>=2.2.3
numpy>=2.1.3
scikit-learn>=1.5.2

# Text Processing
python-dateutil>=2.9.0
pytz>=2024.2
regex>=2024.11.6

# ============================================================================
# SECURITY & CRYPTOGRAPHY
# ============================================================================
cryptography>=43.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
prometheus-client>=0.21.0
python-json-logger>=3.2.1
loguru>=0.7.2
psutil>=5.9.8  # System monitoring

# ============================================================================
# API & NETWORKING
# ============================================================================
httpx>=0.27.2
urllib3>=2.2.3
dnspython>=2.7.0

# ============================================================================
# CONFIGURATION & UTILITIES
# ============================================================================
pyyaml>=6.0.2
toml>=0.10.2
click>=8.1.7
rich>=13.9.4          # Beautiful terminal output
tqdm>=4.67.1          # Progress bars

# ============================================================================
# TESTING & DEVELOPMENT
# ============================================================================
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
black>=24.10.0
flake8>=7.1.1
mypy>=1.13.0

# ============================================================================
# DOCUMENT GENERATION
# ============================================================================
jinja2>=3.1.4
markdown>=3.7
reportlab>=4.2.5      # PDF generation
python-docx>=1.1.2    # Word documents

# ============================================================================
# OPTIONAL: ADVANCED FEATURES
# ============================================================================
# Machine Learning
xgboost>=2.1.2
lightgbm>=4.5.0

# Graph Analysis
networkx>=3.4.2

# Time Series
prophet>=1.1.6

# Geolocation
geoip2>=4.8.1

# ============================================================================
# DEPLOYMENT
# ============================================================================
gunicorn>=23.0.0
supervisor>=4.2.5
